add recorder pages and feature flag
